requirements.txt
psd-tools>=1.9.24
Pillow>=9.0.0
aiofiles>=0.8.0
numpy>=1.21.0 